requirements.txt
boto3>=1.34.0
botocore>=1.34.0
colorama>=0.4.6
tabulate>=0.9.0
requests>=2.31.0
python-dateutil>=2.8.2
rich>=13.7.0
click>=8.1.7
flask>=2.3.0
python-dotenv>=1.0.0

